major refactoring, replacing all the various command line tools with talamel
