+++ src/lib.rs
@@ -30,6 +30,8 @@ extern crate regex;
 #[macro_use]
 extern crate lazy_static;
 
+extern crate talamel;
+
 // extern crate taglib;
 
 // pub mod actions;
