--- src/shelltools/metadata/mod.rs
@@ -1,2 +1,0 @@
-pub mod id3v2;
-pub mod mp4tools;
