+++ src/shelltools/mod.rs
@@ -2,6 +2,5 @@ pub mod bpm;
 pub mod ffmpeg;
 pub mod generic;
 
-pub mod metadata;
 pub mod pipe;
 pub mod sox;
